+++ api/server.go
@@ -91,7 +91,10 @@ type GRPCServer struct {
 	tradeStore         *sqlstore.Trades
 	assetStore         *sqlstore.Assets
 	accountStore       *sqlstore.Accounts
+	rewardStore        *sqlstore.Rewards
 	marketsStore       *sqlstore.Markets
+	delegationsStore   *sqlstore.Delegations
+	epochStore         *sqlstore.Epochs
 
 	eventObserver *eventObserver
 
@@ -139,7 +142,10 @@ func NewGRPCServer(
 	tradeStore *sqlstore.Trades,
 	assertStore *sqlstore.Assets,
 	accountStore *sqlstore.Accounts,
+	rewardStore *sqlstore.Rewards,
 	marketsStore *sqlstore.Markets,
+	delegationStore *sqlstore.Delegations,
+	epochStore *sqlstore.Epochs,
 ) *GRPCServer {
 	// setup logger
 	log = log.Named(namedLogger)
@@ -184,7 +190,10 @@ func NewGRPCServer(
 		tradeStore:              tradeStore,
 		assetStore:              assertStore,
 		accountStore:            accountStore,
+		rewardStore:             rewardStore,
 		marketsStore:            marketsStore,
+		delegationsStore:        delegationStore,
+		epochStore:              epochStore,
 		eventObserver: &eventObserver{
 			log:          log,
 			eventService: eventService,
@@ -335,7 +344,10 @@ func (g *GRPCServer) Start(ctx context.Context, lis net.Listener) error {
 			assetStore:         g.assetStore,
 			accountStore:       g.accountStore,
 			marketDataStore:    g.marketDataStore,
+			rewardStore:        g.rewardStore,
 			marketsStore:       g.marketsStore,
+			delegationStore:    g.delegationsStore,
+			epochStore:         g.epochStore,
 		}
 	} else {
 		g.tradingDataService = tradingDataSvc