+++ cmd/data-node/node/node.go
@@ -110,7 +110,10 @@ type NodeCommand struct {
 	tradeStoreSQL         *sqlstore.Trades
 	networkLimitsStoreSQL *sqlstore.NetworkLimits
 	marketDataStoreSQL    *sqlstore.MarketData
+	rewardStoreSQL        *sqlstore.Rewards
+	delegationStoreSQL    *sqlstore.Delegations
 	marketsStoreSQL       *sqlstore.Markets
+	epochStoreSQL         *sqlstore.Epochs
 
 	vegaCoreServiceClient vegaprotoapi.CoreServiceClient
 
@@ -147,8 +150,11 @@ type NodeCommand struct {
 	networkLimitsSubSQL    *sqlsubscribers.NetworkLimits
 	marketDataSubSQL       *sqlsubscribers.MarketData
 	tradesSubSQL           *sqlsubscribers.TradeSubscriber
+	rewardsSubSQL          *sqlsubscribers.Reward
+	delegationsSubSQL      *sqlsubscribers.Delegation
 	marketCreatedSubSQL    *sqlsubscribers.MarketCreated
 	marketUpdatedSubSQL    *sqlsubscribers.MarketUpdated
+	epochSubSQL            *sqlsubscribers.Epoch
 
 	candleService     *candles.Svc
 	tradeService      *trades.Svc
@@ -338,7 +344,10 @@ func (l *NodeCommand) createGRPCServer(config api.Config, useSQLStores bool) *ap
 		l.tradeStoreSQL,
 		l.assetStoreSQL,
 		l.accountStoreSQL,
+		l.rewardStoreSQL,
 		l.marketsStoreSQL,
+		l.delegationStoreSQL,
+		l.epochStoreSQL,
 	)
 	return grpcServer
 }