+++ cmd/data-node/node/node_pre.go
@@ -130,8 +130,11 @@ func (l *NodeCommand) setupSQLSubscribers() {
 	l.networkLimitsSubSQL = sqlsubscribers.NewNetworkLimitSub(l.ctx, l.networkLimitsStoreSQL, l.Log)
 	l.marketDataSubSQL = sqlsubscribers.NewMarketData(l.marketDataStoreSQL, l.Log, l.conf.SQLStore.Timeout.Duration)
 	l.tradesSubSQL = sqlsubscribers.NewTradesSubscriber(l.tradeStoreSQL, l.Log)
+	l.rewardsSubSQL = sqlsubscribers.NewReward(l.rewardStoreSQL, l.Log)
 	l.marketCreatedSubSQL = sqlsubscribers.NewMarketCreated(l.marketsStoreSQL, l.Log)
 	l.marketUpdatedSubSQL = sqlsubscribers.NewMarketUpdated(l.marketsStoreSQL, l.Log)
+	l.delegationsSubSQL = sqlsubscribers.NewDelegation(l.delegationStoreSQL, l.Log)
+	l.epochSubSQL = sqlsubscribers.NewEpoch(l.epochStoreSQL, l.Log)
 }
 
 func (l *NodeCommand) setupStorages() error {
@@ -167,7 +170,10 @@ func (l *NodeCommand) setupStorages() error {
 		l.networkLimitsStoreSQL = sqlstore.NewNetworkLimits(sqlStore)
 		l.marketDataStoreSQL = sqlstore.NewMarketData(sqlStore)
 		l.tradeStoreSQL = sqlstore.NewTrades(sqlStore)
+		l.rewardStoreSQL = sqlstore.NewRewards(sqlStore)
 		l.marketsStoreSQL = sqlstore.NewMarkets(sqlStore)
+		l.delegationStoreSQL = sqlstore.NewDelegations(sqlStore)
+		l.epochStoreSQL = sqlstore.NewEpochs(sqlStore)
 		l.sqlStore = sqlStore
 	}
 
@@ -244,8 +250,11 @@ func (l *NodeCommand) preRun(_ []string) (err error) {
 			l.networkLimitsSubSQL,
 			l.marketDataSubSQL,
 			l.tradesSubSQL,
+			l.rewardsSubSQL,
+			l.delegationsSubSQL,
 			l.marketCreatedSubSQL,
-			l.marketUpdatedSubSQL)
+			l.marketUpdatedSubSQL,
+			l.epochSubSQL)
 	}
 
 	l.broker, err = broker.New(l.ctx, l.Log, l.conf.Broker, l.chainInfoStore, eventSource)